# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NIAC-Go API
  description: |
    Network In A Can - Go Edition REST API

    This API provides programmatic access to network simulation management,
    device monitoring, configuration updates, and PCAP replay functionality.

    ## Authentication
    All endpoints require Bearer token authentication. Set the token via
    the `NIAC_API_TOKEN` environment variable when starting the server.

    ## CSRF Protection
    State-changing operations (POST, PUT, PATCH, DELETE) require a CSRF token.
    1. Retrieve token from `/api/v1/csrf-token`
    2. Include token in `X-CSRF-Token` header

    ## Rate Limiting
    - Default: 100 requests/second per IP
    - Burst: 200 requests
    - Response: HTTP 429 when exceeded
  version: 2.8.1
  contact:
    name: Kris Armstrong
    url: https://github.com/MustardSeedNetworks/niac-go
  license:
    name: MIT
    url: https://github.com/MustardSeedNetworks/niac-go/blob/main/LICENSE
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://niac.example.com/api/v1
    description: Production server (example)
security:
  - BearerAuth: []
...
paths:
...
  /stats:
    get:
      tags:
        - Statistics
      summary: Get runtime statistics
      description: Returns comprehensive runtime statistics including packet counts, protocol stats, and system info
      operationId: getStats
      responses:
        '200':
          description: Runtime statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistics'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: API token set via NIAC_API_TOKEN environment variable

    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token obtained from /api/v1/csrf-token
  schemas:
    Statistics:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-11-14T10:30:00Z"
        uptime:
          type: string
          example: "2h15m30s"
        version:
          type: string
          example: "2.7.0"
        interface:
          type: string
          example: "eth0"
        device_count:
          type: integer
          example: 50
        goroutines:
          type: integer
          description: Current goroutine count for leak detection
          example: 127
        packets_sent:
          type: integer
          format: int64
          example: 1000000
        packets_received:
          type: integer
          format: int64
          example: 950000
        errors:
          type: integer
          format: int64
          example: 42
        protocol_stats:
          type: object
          properties:
            arp_requests:
              type: integer
              format: int64
              example: 5000
            arp_replies:
              type: integer
              format: int64
              example: 4950
            icmp_requests:
              type: integer
              format: int64
              example: 10000
            icmp_replies:
              type: integer
              format: int64
              example: 9800
            dns_queries:
              type: integer
              format: int64
              example: 2000
            dhcp_requests:
              type: integer
              format: int64
              example: 50
            snmp_queries:
              type: integer
              format: int64
              example: 500
...
    Error:
      type: object
      properties:
        error_code:
          type: string
          example: "invalid_request"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "speed"
              issue:
                type: string
                example: "must be between 0.1 and 10.0"
        request_id:
          type: string
          description: Unique request ID for tracing
          example: "abc123def456789"
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "unauthorized"
            message: "Invalid or missing authentication token"
            request_id: "abc123def456789"
...